Fixed bad output on Tree command when the DB is empty and added ResponseSchema validation layer
